+++ src/features/register/Welcome1Screen/index.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+import Screen from './Welcome1Screen';
+
+export const Welcome1Screen = React.memo(Screen);
